+++ src/browser/src/app/pages/workspace/project/api/env/env.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { EoNgTreeModule } from 'eo-ng-tree';
 import { NzEmptyModule } from 'ng-zorro-antd/empty';
+import { NzTypographyModule } from 'ng-zorro-antd/typography';
 import { SharedModule } from 'pc/browser/src/app/shared/shared.module';
 
 import { EoTableProModule } from '../../../../../modules/eo-ui/table-pro/table-pro.module';
@@ -20,6 +21,7 @@ const COMPONENTA = [EnvListComponent, EnvSelectComponent];
         component: EnvEditComponent
       }
     ]),
+    NzTypographyModule,
     EoNgTreeModule,
     NzEmptyModule,
     SharedModule,