+++ CHANGELOG.md
@@ -1,4 +1,8 @@
+<<<<<<< HEAD
 # [0.4.0](https://github.com/Postcatlab/postcat/compare/v0.2.1...v0.4.0) (2023-03-08)
+=======
+## [0.3.2](https://github.com/Postcatlab/postcat/compare/v0.2.1...v0.3.2) (2023-03-06)
+>>>>>>> main
 
 
 ### Bug Fixes
@@ -22,6 +26,13 @@
 * API Authorization ([05d1077](https://github.com/Postcatlab/postcat/commit/05d1077d3e97c3e9e06afd8a8b15ea02bc1b820e))
 * API Authorization ([938ec7d](https://github.com/Postcatlab/postcat/commit/938ec7d68270c2be4c17cd2e1098377f2db180d2))
 * change description ([0341ba8](https://github.com/Postcatlab/postcat/commit/0341ba8f423cd6554fc83c50bf477dd9d1b36a9c))
+<<<<<<< HEAD
+=======
+* change github project url ([42c801c](https://github.com/Postcatlab/postcat/commit/42c801c49cbd859c876e9aafc2d4bae8fb38b9b6))
+* chat robot ([7878664](https://github.com/Postcatlab/postcat/commit/78786640f464f7370c14aaf14bb224c61a29ae03))
+* delete useless code ([a558094](https://github.com/Postcatlab/postcat/commit/a55809494695b1cdcb44009ca55213ca2fc2a570))
+* delete useless log ([d878582](https://github.com/Postcatlab/postcat/commit/d87858232292f61f89e1412b259dc27b5e644a79))
+>>>>>>> main
 * nps ([426a08f](https://github.com/Postcatlab/postcat/commit/426a08ff1af75d401a8358dae1940f4cfd8f3268))
 * replace reg ([263cb4f](https://github.com/Postcatlab/postcat/commit/263cb4f0d10da3c7cb7ff1bc367541ce159397e8))
 * set registry automaic ([8f64179](https://github.com/Postcatlab/postcat/commit/8f641799d63f77193cea31c2920d590c09f4aaa7))